Removed obsolete interface IRequestHandler. Made ExecuteRequest protected internal: ready for override, but only to be invoked from an invoker which should have more control over the requesthandler lifetime (e.g. disposing when needed, to be implemented).
