--- src/HmLib/Abstractions/IRequestHandler.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace HmLib.Abstractions
-{
-    public interface IRequestHandler
-    {
-        Task<IResponseMessage> HandleRequest(IRequestMessage requestMessage);
-    }
-}
