+++ src/HmLib/Binary/BufferedMessageHandler.cs
@@ -11,7 +11,7 @@ namespace HmLib.Binary
         {
         }
 
-        public override async Task<IResponseMessage> HandleRequest(IRequestMessage requestMessage)
+        internal protected override async Task<IResponseMessage> HandleRequest(IRequestMessage requestMessage)
         {
             try
             {
