+++ src/HmLib/DelegatingRequestHandler.cs
@@ -50,7 +50,7 @@ namespace HmLib
             }
         }
 
-        public override Task<IResponseMessage> HandleRequest(IRequestMessage requestMessage)
+        internal protected override Task<IResponseMessage> HandleRequest(IRequestMessage requestMessage)
         {
             if (requestMessage == null) throw new ArgumentNullException(nameof(requestMessage));
