--- src/HmLib/RequestHandler.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-
-namespace HmLib
-{
-    using Abstractions;
-
-    public abstract class RequestHandler : IRequestHandler, IDisposable
-    {
-        protected RequestHandler() { }
-
-        protected virtual void Dispose(bool disposing) { }
-
-        public void Dispose()
-        {
-            Dispose(true);
-            GC.SuppressFinalize(this);
-        }
-
-        public abstract Task<IResponseMessage> HandleRequest(IRequestMessage requestMessage);
-    }
-}
