+++ src/HmLib/StreamWritingRequestHandler.cs
@@ -18,7 +18,7 @@ namespace HmLib
             _innerStream = innerStream;
         }
 
-        public override async Task<IResponseMessage> HandleRequest(IRequestMessage requestMessage)
+        internal protected override async Task<IResponseMessage> HandleRequest(IRequestMessage requestMessage)
         {
             if (requestMessage == null) throw new ArgumentNullException(nameof(requestMessage));
