+++ src/HmLib/TcpClientRequestHandler.cs
@@ -56,7 +56,7 @@ namespace HmLib
 
         }
 
-        public override async Task<IResponseMessage> HandleRequest(IRequestMessage requestMessage)
+        internal protected override async Task<IResponseMessage> HandleRequest(IRequestMessage requestMessage)
         {
             await ConnectAsync();
